requirements.txt
beautifulsoup4==4.8.0
Click==7.0
Django==2.2.5
django-bootstrap4==1.0.1
django-cors-headers==3.1.1
django-fontawesome-5==1.0.16
django-jsonfield==1.3.1
django-jsonfield-compat==0.4.4
django-simple-history==2.7.3
geobuf==1.1.1
grpcio==1.24.1
grpcio-tools==1.24.1
Pillow==6.2.0
protobuf==3.10.0
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.6
